EET Compatibility
All SoD items already exist in the EET game. However, the mod's other features to add them to creatures or upgrade them by Cespenar/Cromwell would be a welcome addition to EET.

The proposal is to check for EET and skip the item importation part of your mod. The remaining parts work without change.